comments_service/requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Django==3.1.3
factory-boy==3.1.0
Faker==4.16.0
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
